+++ tests/unittests/test_.py
@@ -1,17 +1,17 @@
 """Publick interface correctness tests."""
 import os
-import subprocess
+from subprocess import run
 
 import pytest
-
+# ======================================================================
 SELF_TEST_FLAG = 'LIMEDEV_SELF_TEST'
 SELF_TEST_SKIP = pytest.mark.skipif(SELF_TEST_FLAG in os.environ,
                                     reason = 'Self test')
-
+# ======================================================================
 def runcheck(*args, check = True, **kwargs) -> None:
     """Runs with as default check."""
-    subprocess.run(*args, check = check, **kwargs)
-
+    run(*args, check = check, **kwargs)
+# ======================================================================
 @SELF_TEST_SKIP
 def test_package():
     """Packaging function."""
@@ -21,7 +21,6 @@ def test_package():
 def test_readme():
     """Readme generator."""
     runcheck(('readme'))
-
 # ======================================================================
 @SELF_TEST_SKIP
 class Test_test:
@@ -33,9 +32,13 @@ class Test_test:
         environment[SELF_TEST_FLAG] = ''
         runcheck(('test', 'unittests'), env = environment)
     # ------------------------------------------------------------------
-    def test_profile(self):
+    def test_profiling(self):
         """Profiling handler."""
-        runcheck(('test', 'profiling', '--no_warmup', '--ignore_missing_dot'))
+        runcheck(('test',
+                  'profiling',
+                  '--no_warmup',
+                  '--ignore_missing_dot',
+                  '--function=empty'))
     # ------------------------------------------------------------------
     def test_typing(self):
         """Type check handler."""
@@ -45,7 +48,7 @@ class Test_test:
         """Benchmarking handler."""
         runcheck(('test', 'benchmarking'))
     # ------------------------------------------------------------------
-    def test_lint(self):
+    def test_linting(self):
         """Linting handler."""
         runcheck(('test', 'linting'))
     # ------------------------------------------------------------------